Creating a new go file to improve code reusability
